Minor Fixes and Cleanup
There were a few issues with the running of the demo, and the
installation of the `serve` component, so I wanted to clean those up and
to document the real 'Getting Started' steps so that folks didn't have
to go through what I went through. At the same time, I learned a lot
about how the zoids are built, and the locations of definitions, etc.
This is a good demo now, and I'm glad to have figured it all out.