Added more to README for help
We are really starting to flesh-out some of the help that I would have
found very helpful when I was starting this up. This is nothing that an
experienced developer in Node might not already know, but it's news to
me, and so I assume it might be news to others.