Added the CHANGELOG and tools
We really needed to have a running CHANGELOG.md for the base project,
and we then needed to be able to easily bump the version number - major,
minor, or patch - and it was easy enough to use the `npm` tools in the
scripts section of the 'package.json' file, and wire it in there. We
also created the CHANGELOG.md file, and updated it for these additions.